remove DefaultCache and DefaultLogger from Beagle
