+++ iOS/Example/BeagleDemo/BeagleDemo/Cache/LRU/DoublyLinkedList.swift
