+++ iOS/Example/BeagleDemo/BeagleDemo/Cache/Tests/CacheManagerDefaultTests.swift
