+++ iOS/Example/BeagleDemo/BeagleDemo/Network/HttpRequestBuilder.swift
